Split bufr4.py into bufr4.py and bufr2geojson.py
