+++ config_examples/data-mappings.yml.example-synop-bufr
@@ -8,7 +8,7 @@ data:
                     - ${WIS2BOX_STORAGE_INCOMING}
                   file-pattern: '*'
             bufr4:
-                - plugin: wis2box.data.bufr4.ObservationDataBUFR2GeoJSON
+                - plugin: wis2box.data.bufr2geojson.ObservationDataBUFR2GeoJSON
                   buckets:
                     - ${WIS2BOX_STORAGE_PUBLIC}
                   file-pattern: '^WIGOS_(\d-\d+-\d+-\w+)_.*\.bufr4$'
