+++ wis2box/resources/data-mappings.yml
@@ -7,11 +7,8 @@ data:
                   notify: true
                   file-pattern: '^WIGOS_(\d-\d+-\d+-\w+)_.*\.csv$'
             bufr4:
-                - plugin: wis2box.data.bufr4.ObservationDataBUFR2GeoJSON
+                - plugin: wis2box.data.bufr2geojson.ObservationDataBUFR2GeoJSON
                   file-pattern: '^WIGOS_(\d-\d+-\d+-\w+)_.*\.bufr4$'
-            geojson:
-                - plugin: wis2box.data.geojson.ObservationDataGeoJSON
-                  file-pattern: '^WIGOS_(\d-\d+-\d+-\w+)_.*\.geojson$'
     ita.roma_met_centre.data.core.weather.surface-based-observations.SYNOP:
         plugins:
             bin:
@@ -19,11 +16,8 @@ data:
                   notify: true
                   file-pattern: '^.*\.bin$'
             bufr4:
-                - plugin: wis2box.data.bufr4.ObservationDataBUFR2GeoJSON
+                - plugin: wis2box.data.bufr2geojson.ObservationDataBUFR2GeoJSON
                   file-pattern: '^WIGOS_(\d-\d+-\d+-\w+)_.*\.bufr4$'
-            geojson:
-                - plugin: wis2box.data.geojson.ObservationDataGeoJSON
-                  file-pattern: '^WIGOS_(\d-\d+-\d+-\w+)_.*\.geojson$'
     dza.alger_met_centre.data.core.weather.surface-based-observations.SYNOP:
         plugins:
             bufr4:
@@ -32,10 +26,7 @@ data:
                   buckets:
                     - ${WIS2BOX_STORAGE_INCOMING}
                   file-pattern: '^.*\.bufr4$'
-                - plugin: wis2box.data.bufr4.ObservationDataBUFR2GeoJSON
+                - plugin: wis2box.data.bufr2geojson.ObservationDataBUFR2GeoJSON
                   buckets:
                     - ${WIS2BOX_STORAGE_PUBLIC}
                   file-pattern: '^WIGOS_(\d-\d+-\d+-\w+)_.*\.bufr4$'
-            geojson:
-                - plugin: wis2box.data.geojson.ObservationDataGeoJSON
-                  file-pattern: '^WIGOS_(\d-\d+-\d+-\w+)_.*\.geojson$'
